refactoring, added cart clearing after order creation, added pageable, added status updating endpoint

+++ src/main/java/com/example/onlinebookstore/dto/shoppingcart/CartItemDto.java
@@ -5,5 +5,5 @@ import jakarta.validation.constraints.Positive;
 
 public record CartItemDto(
         @Positive @NotNull Long bookId,
-        @Positive @NotNull Long quantity) {
+        @Positive @NotNull int quantity) {
 }
